# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Clinic",
        "description": "API پروژه ",
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/appointments": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "دریافت لیست تمام نوبت‌ها (فقط ادمین)",
                "description": "این API لیست تمام نوبت‌ها را فقط در صورتی که کاربر ادمین باشد برمی‌گرداند",
                "operationId": "getAllAppointments",
                "responses": {
                    "200": {
                        "description": "لیست نوبت‌ها با موفقیت دریافت شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "provider_id": {
                                            "type": "integer",
                                            "example": "دکتر محمد صادقی"
                                        },
                                        "service_id": {
                                            "type": "integer",
                                            "example": "ویزیت دکتر عمومی"
                                        },
                                        "customer_id": {
                                            "type": "integer",
                                            "example": "حسین ملکی"
                                        },
                                        "appointment_date": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-01-10"
                                        },
                                        "start_time": {
                                            "type": "string",
                                            "example": "09:00"
                                        },
                                        "end_time": {
                                            "type": "string",
                                            "example": "09:30"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "completed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "عدم احراز هویت"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/reports/daily": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "گزارش نوبت‌ها (فقط ادمین)",
                "description": "دریافت گزارش نوبت‌ها به صورت مرتب‌شده بر اساس تاریخ نوبت",
                "operationId": "getAppointmentsReport",
                "responses": {
                    "200": {
                        "description": "گزارش نوبت‌ها با موفقیت دریافت شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "appointment_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "1404-1-15"
                                            },
                                            "data": {
                                                "properties": {
                                                    "provider_id": {
                                                        "type": "string",
                                                        "example": "دکتر محمدی"
                                                    },
                                                    "service_id": {
                                                        "type": "string",
                                                        "example": "ویزیت حلق و بینی"
                                                    },
                                                    "customer_id": {
                                                        "type": "string",
                                                        "example": "علی احمدی"
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "example": "09:00"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "example": "09:30"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "reserved"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "عدم دسترسی (کاربر ادمین نیست)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "عدم احراز هویت"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "description": "Bearer {token}"
            }
        }
    }
}